// components/interfaces/Events.ts
export default interface EventsProps {
  idEvent: string;
  idSoccerXML: string;
  idAPIfootball: string;
  strEvent: string;
  strEventAlternate: string;
  strFilename: string;
  strSport: string;
  idLeague: string;
  strLeague: string;
  strSeason: string;
  strDescriptionEN: string;
  strHomeTeam: string;
  strAwayTeam: string;
  intHomeScore: string;
  intRound: string;
  intAwayScore: string;
  intSpectators: string;
  strHomeGoalDetails: string;
  strHomeRedCards: string;
  strHomeYellowCards: string;
  strHomeLineupGoalkeeper: string;
  strHomeLineupDefense: string;
  strHomeLineupMidfield: string;
  strHomeLineupForward: string;
  strHomeLineupSubstitutes: string;
  strHomeFormation: string;
  strAwayRedCards: string;
  strAwayYellowCards: string;
  strAwayGoalDetails: string;
  strAwayLineupGoalkeeper: string;
  strAwayLineupDefense: string;
  strAwayLineupMidfield: string;
  strAwayLineupForward: string;
  strAwayLineupSubstitutes: string;
  strAwayFormation: string;
  intHomeShots: string;
  intAwayShots: string;
  dateEvent: string;
  dateEventLocal: string;
  strDate: string;
  strTime: string;
  strTimeLocal: string;
  strTVStation: string;
  idHomeTeam: string;
  idAwayTeam: string;
  strResult: string;
  strCircuit: string;
  strCountry: string;
  strCity: string;
  strPoster: string;
  strFanart: string;
  strThumb: string;
  strBanner: string;
  strMap: string;
  strTweet1: string;
  strTweet2: string;
  strTweet3: string;
  strVideo: string;
  strPostponed: string;
  strLocked: string;
}